+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this repository are documented here. We are using [Semantic Versioning for Documents](https://semverdoc.org/), in which a version number has the format `major.minor.patch`.
 
+## 2.3.0 - 2020-12-02
+
+- Added scripts for multi-lifetime evaluation
+- Added validation for scenario info (complexity/difficulty)
+- Added type hints
+- Updated reference to metrics specification document 0.65
+- Updated calculation for performance recovery to use negative slope
+
 ## 2.2.0 - 2020-11-19
 
 - Added validation to performance maintenance
@@ -28,7 +36,7 @@ All notable changes to this repository are documented here. We are using [Semant
 - Updated initial metrics to include all major identified lifelong learning metrics
 - Removed dependence on learnkit
 - Refectored metrics calculations to integrate with different log format
-- Updated workflow for storing single task expert data
+- Updated workflow for storing single-task expert data
 - Updated READMEs and example scripts
 - Added new example logs along with their metrics outputs